requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
biopython>=1.79
torch>=1.9.0
mdtraj>=1.9.7
pymol-open-source>=2.5.0
scikit-learn>=1.0.0
seaborn>=0.11.0
tqdm>=4.62.0
argparse
yaml>=6.0
h5py>=3.7.0
networkx>=2.6.0 